deleting hibernate / simulation business + modif jdbc

+++ src/business/excursion/ElementTarifiable.java
@@ -1,7 +1,9 @@
 package business.excursion;
 
+import java.math.BigDecimal;
+
 public interface ElementTarifiable {
 
-	double getPrix();
-	
+    BigDecimal getPrix();
+
 }
